+++ plone/app/event/tests/test_event_listing.py
@@ -53,8 +53,9 @@ class TestEventsListingDX(AbstractSampleDataEvents):
         headers, output, request = make_fake_response(self.request)
         view = self.portal.restrictedTraverse('@@event_listing_ical')
         view()
-        self.assertEqual(len(headers), 2)
+        self.assertEqual(len(headers), 3)
         self.assertEqual(headers['Content-Type'], 'text/calendar')
+        self.assertTrue('Content-Length' in headers)
         icalstr = ''.join(output)
         self.assertTrue('Long Event' in icalstr)